add test for unmarshal_json add unmarshal_json for the data_part
