+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing to a2a-go
+
+## pull request 
+
+you can edit the source code by cloning this repository
+
+```shell
+git clone https://github.com/yeeaiclub/a2a-go.git
+```
