+++ sdk/types/event.go
@@ -27,136 +27,6 @@ type Event interface {
 	Done() bool
 }
 
-type Role string
-
-const (
-	Agent Role = "agent"
-	User  Role = "user"
-)
-
-// Message Represents a single message exchanged between user and agent
-type Message struct {
-	Role             Role           `json:"role"`
-	TaskID           string         `json:"task_id,omitempty"`
-	ContextID        string         `json:"context_id,omitempty"`
-	Extensions       []string       `json:"extensions,omitempty"`
-	Kind             string         `json:"kind,omitempty"`
-	MessageID        string         `json:"message_id,omitempty"`
-	ReferenceTaskIDs []string       `json:"referenceTaskIds,omitempty"`
-	Parts            []Part         `json:"parts,omitempty"`
-	Metadata         map[string]any `json:"metadata,omitempty"`
-}
-
-func (m *Message) Done() bool {
-	return true
-}
-
-func (m *Message) GetContextId() string {
-	return m.ContextID
-}
-
-func (m *Message) GetTaskId() string {
-	return m.TaskID
-}
-
-func (m *Message) EventType() string {
-	return "message"
-}
-
-func (m *Message) UnmarshalJSON(data []byte) error {
-	type Alias Message // avoid recursion
-	aux := &struct {
-		Parts []json.RawMessage `json:"parts,omitempty"`
-		*Alias
-	}{
-		Alias: (*Alias)(m),
-	}
-	if err := json.Unmarshal(data, &aux); err != nil {
-		return err
-	}
-	m.Parts = nil
-	for _, raw := range aux.Parts {
-		var kindHolder struct {
-			Kind string `json:"kind"`
-		}
-		if err := json.Unmarshal(raw, &kindHolder); err != nil {
-			return err
-		}
-
-		var part Part
-		switch kindHolder.Kind {
-		case "text":
-			var tp TextPart
-			if err := json.Unmarshal(raw, &tp); err != nil {
-				return err
-			}
-			part = &tp
-		case "data":
-			var dp DataPart
-			if err := json.Unmarshal(raw, &dp); err != nil {
-				return err
-			}
-			part = &dp
-		case "file":
-			var fp FilePart
-			if err := json.Unmarshal(raw, &fp); err != nil {
-				return err
-			}
-			part = &fp
-		default:
-			return fmt.Errorf("unknown part kind: %q", kindHolder.Kind)
-		}
-		m.Parts = append(m.Parts, part)
-	}
-	return nil
-}
-
-type TaskStatus struct {
-	Message   *Message  `json:"id,omitempty"`
-	State     TaskState `json:"state"`
-	TimeStamp string    `json:"time_stamp,omitempty"`
-}
-
-type Task struct {
-	Id        string         `json:"id"`
-	ContextId string         `json:"context_id"`
-	History   []*Message     `json:"history,omitempty"`
-	Kind      string         `json:"kind,omitempty"`
-	Status    TaskStatus     `json:"task_status,omitempty"`
-	Metadata  map[string]any `json:"metadata,omitempty"`
-	Artifacts []Artifact     `json:"artifacts,omitempty"`
-}
-
-type Artifact struct {
-	ArtifactId  string         `json:"artifact_id,omitempty"`
-	Description string         `json:"description,omitempty"`
-	Extensions  []string       `json:"extension,omitempty"`
-	Metadata    map[string]any `json:"metadata,omitempty"`
-	Name        string         `json:"name,omitempty"`
-	Parts       []Part         `json:"parts,omitempty"`
-}
-
-func (t *Task) Done() bool {
-	return t.Status.State == COMPLETED ||
-		t.Status.State == CANCELED ||
-		t.Status.State == FAILED ||
-		t.Status.State == REJECTED ||
-		t.Status.State == UNKNOWN ||
-		t.Status.State == InputRequired
-}
-
-func (t *Task) GetContextId() string {
-	return t.ContextId
-}
-
-func (t *Task) GetTaskId() string {
-	return t.Id
-}
-
-func (t *Task) EventType() string {
-	return "task"
-}
-
 // TaskArtifactUpdateEvent Send by server during sendStream and subscribe requests
 type TaskArtifactUpdateEvent struct {
 	TaskId    string         `json:"task_id"`
@@ -264,17 +134,3 @@ func (s *StreamEvent) EncodeJSONRPC(encoder *json.Encoder, id string) error {
 	}
 	return nil
 }
-
-type TaskState string
-
-const (
-	SUBMITTED     = "submitted"
-	WORKING       = "working"
-	InputRequired = "required"
-	COMPLETED     = "completed"
-	CANCELED      = "canceled"
-	FAILED        = "failed"
-	REJECTED      = "rejected"
-	AUTH_REQUIRED = "auth_required"
-	UNKNOWN       = "unknown"
-)
